bpf-snoops: change bpf events data-types
o Do not use linux/types.h (especially in user-space part).
o Do not enforce u64 for pointers, instead use arch-specific
  unsigned long.
o Do not enforce u64 size

Also re-arrange includes (cros format), except vmlinux.h
(this one needs to be before bpf includes).

It's almost impossible to make the linters happy, because linters
don't know that we share data-structures between the "target BPF"
compiled kernel code and C++ code. Yet it seems that as of now
the linters are totally OK with our code (including cros format).

BUG=b:353909098
TEST=emerge-brya-kernelnext bpf-snoops

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5721931
Tested-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Commit-Queue: Sergey Senozhatsky <senozhatsky@chromium.org>

+++ bpf-snoops/include/libsnoop.h
@@ -5,12 +5,14 @@
 #ifndef BPF_SNOOPS_INCLUDE_LIBSNOOP_H_
 #define BPF_SNOOPS_INCLUDE_LIBSNOOP_H_
 
+#include <errno.h>
+#include <stdint.h>
+#include <stdio.h>
+
+#include <string>
+
 #include <bpf/bpf.h>
 #include <bpf/libbpf.h>
-
-#include <errno.h>
-#include <stdio.h>
-#include <string>
 
 namespace libsnoop {
 
@@ -49,8 +51,8 @@ namespace libsnoop {
 int init_stack_decoder(void);
 void release_stack_decoder(void);
 
-void decode_ustack(__u32 pid, __u64* ents, __s32 num_ents);
-void decode_kstack(__u64* ents, __s32 num_ents);
+void decode_ustack(uint32_t pid, uintptr_t* ents, uint32_t num_ents);
+void decode_kstack(uintptr_t* ents, uint32_t num_ents);
 
 int lookup_lib(const char* name, std::string& path);
